Fix incorrect assumption for package manager error codes
I incorrectly assumed I was getting the error code, I was not.
God I hate shell. To get the error code, you must run the command
and access the special variable $?, which will return the error
code of the last run command. We will use this as check instead.

Also, error code for brew/apt-get returns 1 when it exists, so thats
fun.