Make shell config process less complicated
I was being real dumb with the adding of $HOME/bin to the PATH
variable. Instead of caring about if bin is in the path variable
and exporting the PATH variable in .bash_profile, I can just add a check
to see if $HOME/bin exists in .bashrc, and if it does,
I append it to the PATH variable. Since .bash_profile sources
.bashrc as well, this should append $HOME/bin to the path variable
everytime I open a shell.

Move the creation of $HOME/bin to the first thing we do in the script,
so it will exist when we go to setup our shell and check to see if
it exists so we can add it to the path variable