Fix word splitting/quoting issue for dash
Turns out in certain versions of dash(0.5.8-2.10)
you cannot do an export and assignment of a variable
in cases such as command substitution:
        f() { local e=$1; }
        export g=$(printf '%s' "foo 2")

The fix to this is either quoting the right hand of the assignment
or seperating the assignment and exporting of the variable into
two seperate lines.

We are doing the latter, as this also fixes masking return values that
happens with exporting and assigning a variable on the same line.
Instead of returning the exit code of the cmd being assigned to a variable,
it will return the exit code of export, which is always true.

See shellcheck SC2155 for more info.