Setup logic to allow common aliases between shells
Created a common folder in the shell folder to store whatever
is common between shells like aliases, functions etc.

The aliases are in a seperate file to make it easier to seperate/change
if needed. Same thing should apply to anything else I add in the common
folder.

Add logic to .bashrc to source the alias file during startup