Add logic to skip installing already installed packages
Did a check to to see if a package is installed and if
it is,skip it automatically.

Package names used to query if its installed is the folder name.
Renamed the openssh and ctags folders to the correct package
name that the package manager expects. Not sure how
this worked before and now it doesn't.

For now the package manager for the query logic
is hard coded to dpkg(debian). We need to
create a function to dymanically choose the package manager
similar to the _get_install_command function.