Fix _dir_files_total_line_count to work with non GNU find
For GNU find, if you don't specify what path you want to search in,
it will default to the current path the command is run in(.).

However, for systems that don't use GNU implementations of commands
by default such as BSD, find requires you to explicitly tell it what path
to start at. The simple fix was to just add a . after the find
keyword