Refactor export/alias seperation logic
Refactor the scripts to create a .export and .alias folder. This is
where all the the export scripts for apps will go. The aliases/exports
that are common across all shells will be at the root of the folders
and is called common-export and common-alias. Made a common-config
file that will be used to put all the common configuration logic
I want for various files.