Add setup to install more software for personal use
Now that I'm trying to fully go to linux, I'm trying
to replicate the setup I have on Windows. Main things
in this commit are keepassxc and dropbox. Python
is needed for the dropbox "CLI" which is really
a python script.

Needed to include setup for nvidia drivers since those
don't come installed out of the box for linux