Alter package check command to use which
Using the which command is a little cleaner to determine what
package manager to use as I don't need to try and run
a specific package manager command and see if it returns
an error code of 0. With this, if which returns 0,
I know that the package manager exists on the machine
and I can use it.