Fix dropbox/qdirstat for pacman install
Dropbox/qdirstat are part of the AUR repo.
Bascially, a community run first of the offical
arch repo. To pull from the AUR, you need
to pull down the tar file, extract it,
and run the makepkg command to install it
via pacman. Altered their scripts to do just
that