Split package manager function into two in software setup.sh
I always want to prioritize native package installation(pacman, dpkg)
over non native package installtion(brew, flatpak, etc)

I split the get_package_manager function into two functions,
get_native_package_manager and get_non_native_package_manager
Then, I am checking to see if I could find a software install script
for the native package manager. If so we install it.

If native doesn't exist, we check to see
if we can find an install script for a non native package manager.

If we can't find either, we check to see if we can find an install
script that doesn't depend on a package manager.

If none of these types of scripts are found, we don't install the
software