Add export setup for vim
Add an export for vim to set the EDITOR
env variable to vim.

This can be used by sudo -e or sudoedit
to edit a file as sudo.

It copys the file to the /tmp directory, lets you
make your changes, then moves the changed file back
to the original location

This command is useful because it protects
against exploits related to your editor

It should also respect your editors config files