Change location of bin,.export,and .alias folders
It was kinda messy having the .export,.alias, and bin folders
in there own location under $HOME. Put them all
under .distro-config so we know what came from the install
process. Change scripts to point to .distro-config for
copying of scripts, exports, etc.

TODO Still need to find a way for the install location
for bin .export and .alias to be configurable