Refactor installation of aliases, bin scripts, exports
Most of the scripts for aliases, bin scripts, and exports
in each software folder were doing the same thing.
So I made a script for each type and will have the setup
script call that instead of the scripts that were
in the software folder. This makes it easier
to change the config files location in the future
and gets of redundant scripts.