Seperate out install config logic to own script
To be consistant with what I did for exports, bin scripts,
and aliases, the process to install configs will be
refactored out to its own script.

Sadly, we can't get rid of the install-config scripts
that are in the software dirs since the location of each
softwares configs are so different.

However, if we want to only install configs for a piece of
software for any reason, this script will make it possible
to do so.