Setup cron files for github download repo scripts
With certain repos getting taken down(Yuzu, Citra), I need
to be as up to date as possible in case that happens. So
I made crontab files that will be manually installed if I want to
do this.

Changed setup process for software to put the cron files
in .distro-config/cron just so I know where I can find them.

Any custom scripts I have made should be ok to use in the crontab
since the bin folder in .distro-config should be on the PATH