Change rm confirmation to be less verbose
rm -i prompts for every single file deletion recursively. This is
a giant pain in the ass for deleting large amounts of files. Instead
we should use -I(--interactive=once), which prompts when more than
3 files are deleted or a directory is deleted. This will have the
same outcome of making rm more safe for us but without being  so
verbose. Also added the --verbose flag so we can see what rm is
explicitly deleting.