Remove subshells and replace with dirname/basename
I had so many subshells in this project and scripts
that depended on being in specific dirs to work.

I replaced the subshells with dirname and basename to
make it so you can run any part of the setup process
under any directory(inside or out of the project) and still be ok.