Remove .profiles and replace with software folder name
I wasn't really using the .profiles and it was kinda a pain
in the ass to manage. So from now on the way to specify specific
software to install is the --software flag and the values are
the folders under software.

If nothing is provided for --software, it will prompt you to
install everything