Change lgogdownloader script to list out each game
Initially, the lgogdownloader-download-all-games.sh
script did download all the games in my GOG library.

However, some games are so bloated with the game
size and patches that I had to not have them so
it doesn't take up all my dropbox space.

There is no good way to exclude games in lgogdownloader,
so I had to list out every game I wanted and exclude the
ones I don't. The ones I did not want are commented out
in the beginning of the script in case I change my
mind.

This also means that if I buy a game and want the script
to download it, I need to manually list it in this script.

Lgogdownloader is very annoying.