Refactor distro folder to be set in config file
The distro folder is where all the scripts, aliases, exports, etc
that are managed by this repo live. Previously, this location was
set in various scripts to the point where changing it would require
touching multiple files.

This change introduces the distro-config.sh file at the top of the
repo. This script defines all the environment variables and checks
that were done across multiple scripts. If I want to change anything,
it is now in one location instead of multiple scripts.

distro-config.sh is used in the following way:
        Any scripts used for setting up things in the repo that
        need to move something to the distro folder will call
        distro-config.sh to find out the locations it needs.

        During shell setup, the distro-config.sh file will be copied
        to the distro folder itself, for usage by the shell config
        files (.bashrc, .profile, etc). This is because to make
        all the scripts, aliases, exports, etc avaliable in the shell,
        we need to source them. To source them, we need to know their
        location, which is the distro folder configured in
        distro-config.sh

This is better than it was but its not perfect. The shell configs
(.bashrc, .profile, etc) still directly references the distro folder
directly, because it needs to source distro-config.sh to know
all the various locations to look to source things. This
is done in shell/common/common-config.sh right now.

With this flaw, we would have to make two changes to change the
distro folder. One in distro-config.sh and in
shell/common/common-config.sh.

Still trying to figure out how to fix that, might need to dynamically
create the shell configs instead.

Still a vast improvment over what it was though