Remove hard coded distro directory path in common-config.sh
See previous commit for a more in depth reason why.

But mainly, its so we aren't referencing the distro directory,
which can be changed dynamically in the shell configs (.bashrc,
.profie, etc)

Now, during shell setup we generate a .distro-boostrap file.
This file lives in the users HOME directory, and contains
the location of the distro directory path. This will be
regenerated everytime setup is run, so its updated automatically
if the distro folder changes.

Then, the shell config references the distro boostrap file for
the distro directory location so it can load distro-config.sh

The shell config has sane defaults for the distro directory
location if the .distro-boostrap file doesn't exist(.distro).