Remove ** and replace with find command for POSIX compliance
**(double glob or globstar) is a bash specific feature that that matches
files and directories recursively.

In shell/common/common-config.sh, we are using this to recursively
find program specific exports / aliases that were installed by this
repo. This is so we can source them so they are usable in our shell.

Since we want to be POSIX compliant I have replaced this with using
the find command instead. This will have the same behavior and
recursively find the alias and export files for sourcing